--- src/components/Messages/Companion/Companion.module.css
@@ -1,22 +1,0 @@
-.companions {
-	grid-area: c;
-}
-.companions_items  {
-	text-decoration: none;
-}
-.companion {
-	width: 80%;
-	margin: 20px auto;
-	text-align:center;
-	background: lightskyblue;
-	border: 1px solid whitesmoke;
-	height: 45px;
-	border-radius: 20%;
-	position: relative;
-}
-.companion p {
-	padding:0;
-	margin:10px auto;
-	font-size: 20px;
-	color: white;
-}
