--- src/components/Messages/Dialogs/Dialogs.module.css
@@ -1,3 +1,0 @@
-.dialogs {
-	grid-area: d;
-}
